make sure viewport is initialized
It's illegal to pass a render-area with a width or height of zero to
Vulkan. But we initialize the viewport to zero, which prevents us from
rendering the initial frames.

So let's make sure we initialize the viewport to the full window size.
This matches OpenGL's behavior.

This fixes validation errors and missing rendering on start-up.